Minor Database Consolidation

+++ database/migrations/2021_11_01_150705_create_order_customer_adjustments_table.php
@@ -16,7 +16,7 @@ class CreateOrderCustomerAdjustmentsTable extends Migration
         Schema::create('order_customer_adjustments', function (Blueprint $table) {
             $table->id();
             $table->foreignId('order_customer_id')->constrained()->onDelete('cascade');
-            $table->float('amount');
+            $table->decimal('amount', 12);
             $table->text('reason');
             $table->date('date');
             $table->softDeletes();
